renamed algebra to algebras
