+++ setup.py
@@ -70,7 +70,7 @@ if sys.version_info[:2] < (2, 7):
 # Check that this list is uptodate against the result of the command:
 # python bin/generate_module_list.py
 modules = [
-    'sympy.algebra',
+    'sympy.algebras',
     'sympy.assumptions',
     'sympy.assumptions.handlers',
     'sympy.benchmarks',
@@ -268,6 +268,7 @@ class run_benchmarks(Command):
 # Check that this list is uptodate against the result of the command:
 # python bin/generate_test_list.py
 tests = [
+    'sympy.algebras.tests',
     'sympy.assumptions.tests',
     'sympy.calculus.tests',
     'sympy.categories.tests',
