+++ sympy/__init__.py
@@ -73,7 +73,7 @@ from .integrals import *
 from .tensor import *
 from .parsing import *
 from .calculus import *
-from .algebra import *
+from .algebras import *
 # Adds about .04-.05 seconds of import time
 # from combinatorics import *
 # This module is slow to import:
